Adding MUI, ESLint and ESLint configuration
