+++ pages/api/hello.js
@@ -1,0 +1,5 @@
+export default async (req, res) => {
+  res.status(200).json({
+    message: "Hello World!",
+  });
+};
